refactor(types): extract inline types to centralized type definitions
- Create `$lib/types/events.ts` with `K8sInvolvedObject` and `K8sEvent`
- Create `$lib/types/reconciliation.ts` with `ReconciliationEntry`
- Refactor `$lib/types/resource.ts` to re-export from new files
- Remove duplicate `K8sEvent` from server `events.ts`, import shared type
- Import `HealthCheckResult`/`ClusterHealthCheck` from `$lib/server/clusters`
  in admin clusters page instead of redefining inline

Closes #201
